NamasteKart/requirements.txt
pandas
secure-smtplib
python-decouple